+++ test/runtests.jl
@@ -38,3 +38,5 @@ join_db2 = DB.db_table(db, df3, "df_join2");
 @testset "TidierDB to TidierData comparisons" verbose = true begin
    include("comp_tests.jl")
 end
+
+DB.DBInterface.close(db)